basic html form Just created a simple demo html page to test sending POST requests to the server.
Need to figure out what to send the user back, and how to handle that data,
but it does work with sending POST req's to the server itself.

Also need to work on figuring out the purpose of this site/server,
as well as what kind of tools I want to impliment (sql, php, javascript, etc).